server/requirements-simple.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
google-auth==2.40.3
PyJWT==2.10.1
requests==2.32.5
python-dateutil==2.9.0.post0

